Fixed recipients
Assigning roles from the start is incompatible with how the data is organized in the database. Recipients need to exist before they can be assigned roles. And a recipient without a role is the default state. Hence I removed the option to specify roles from the start. Easier fix than making two consecutive requests and managing errors between two states.